+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+1.0.0 / 2016-05-01
+==================
+
+  * api stable; releasing 1.0
+  * engine.run() now returns triggered events
+
 1.0.0-beta10 / 2016-04-16
 ==================